color config and beginnings of gauge config logic
Started in on the gauge configuration logic on a per page basis with
some generic functions to access sensor data.  Decided to do if-then
decoding of color picking (it's less mess).  Decided to use a separate
entry to determine pin per sensor.  Also decided that "0" would
designate obd II sensor.